requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pinecone-client
sentence-transformers
pydantic
python-multipart
pandas
numpy
pytest
httpx
aiohttp